Refactor URLs into existing constants file.
