+++ src/app/constants.ts
@@ -6,3 +6,7 @@ export enum AccountType {
   GROUP_IRONMAN = 'GROUP_IRONMAN',
   HARDCORE_GROUP_IRONMAN = 'HARDCORE_GROUP_IRONMAN'
 }
+
+export const discordUrl = 'https://discord.gg/loghunters';
+export const gitHubUrl = 'https://github.com/evansloan/collectionlog.net';
+export const pluginUrl = 'https://runelite.net/plugin-hub/show/collection-log';
