--- src/constants/Urls.ts
@@ -1,5 +1,0 @@
-export const discordUrl = 'https://discord.gg/loghunters';
-
-export const gitHubUrl = 'https://github.com/evansloan/collectionlog.net';
-
-export const pluginUrl = 'https://runelite.net/plugin-hub/show/collection-log';
